requirements.txt
apache-airflow~=2.10.2
apache-airflow-providers-postgres~=5.13.1